do not export fmtProcessEvents
